menyelasaikan sistem frontend dan backend, (masih ada defect ketika bebrapa pelanggan melakukan pemesanan)

+++ database/migrations/2025_05_16_154318_create_order_items_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->foreignId('pesanan_id')->constrained('orders')->onDelete('cascade');
             $table->foreignId('menu_id')->constrained('menus')->onDelete('cascade');
             $table->integer('jumlah');
-            $table->decimal('subtotal', 10, 2);
+            $table->bigInteger('subtotal');
             $table->timestamps();
         });
     }
